requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
transformers==4.55.4
torch==2.8.0+cu126
torchvision==0.23.0+cu126
torchaudio==2.8.0+cu126
accelerate==1.10.1
datasets==4.0.0
sentencepiece==0.2.1
protobuf==5.29.5
py-cpuinfo==9.0.0
psutil==5.9.5
scipy==1.16.1
safetensors==0.6.2
einops==0.8.1
peft==0.17.1
tqdm==4.67.1
omegaconf==2.3.0
gradio==5.43.1
numpy==2.0.2
pandas==2.2.2
scikit-learn==1.6.1
matplotlib==3.10.0
seaborn==0.13.2
tokenizers==0.21.4
fastapi==0.115.6
uvicorn==0.29.0
pydantic==2.10.3
PyYAML==6.0.2
jsonschema==4.22.0
